requirements.txt
openai>=1.30.0
chromadb>=0.5.3
python-dotenv>=1.0.1
rich>=13.7.1
